+++ examples/external_kms_no_policy/main.tf
@@ -33,7 +33,6 @@ module "external_kms_no_policy" {
   types             = ["String", "SecureString", "StringList"]
   values            = ["foo was here", "bar was here", "baz was here"]
   kms_key_create    = false
-  kms_key_id        = aws_kms_key.this.id
   kms_key_arn       = aws_kms_key.this.arn
   iam_policy_create = true