+++ main.tf
@@ -10,7 +10,7 @@ resource "aws_ssm_parameter" "overwrite" {
   type        = element(var.types, count.index)
   value       = element(var.values, count.index)
 
-  key_id          = element(var.types, count.index) == "SecureString" ? var.kms_key_create ? element(concat(aws_kms_key.this.*.id, [""]), 0) : var.kms_key_id != "" ? var.kms_key_id : null : null
+  key_id          = element(var.types, count.index) == "SecureString" ? var.kms_key_create ? element(concat(aws_kms_key.this.*.arn, [""]), 0) : var.kms_key_arn != "" ? var.kms_key_arn : null : null
   overwrite       = true
   allowed_pattern = element(concat(var.allowed_patterns, [""]), count.index)
 
@@ -30,7 +30,7 @@ resource "aws_ssm_parameter" "no_overwrite" {
   type        = element(var.types, count.index)
   value       = element(var.values, count.index)
 
-  key_id          = element(var.types, count.index) == "SecureString" ? var.kms_key_create ? element(concat(aws_kms_key.this.*.id, [""]), 0) : var.kms_key_id != "" ? var.kms_key_id : null : null
+  key_id          = element(var.types, count.index) == "SecureString" ? var.kms_key_create ? element(concat(aws_kms_key.this.*.arn, [""]), 0) : var.kms_key_arn != "" ? var.kms_key_arn : null : null
   allowed_pattern = element(concat(var.allowed_patterns, [""]), count.index)
 
   lifecycle {