+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.0.0
 
 * fix: Add `parameters_count` variable. You must update you module with this variable, otherwise, no SSM paramters will be created. This fix a terraform 0.12 pre-processor issue.
+* tech: Remove `kms_key_id` variable. SSM parameter resource can now accept both KMS key ID or KMS key ARN.
 
 ## 1.1.0